경매장 완성..? 디자인 수정 조금 필요

+++ app/src/main/java/com/example/myapplication/dataModel/marketsItemData.kt
@@ -11,6 +11,6 @@ data class marketsItemData(
     )
 data class ItemStats(
     @SerializedName("Date") val date: String?,
-    @SerializedName("AvgPrice") val avgPrice: Int?,
+    @SerializedName("AvgPrice") val avgPrice: Double?,
     @SerializedName("TradeCount") val tradeCount: Int?,
 )
